New files within data files
Sorry for the large amount of changes. I made separate files for the ACA and RSVP data within the data folder so it would be easier for me to sort through. The minor changes in ACAAnalysisNew were to tailor the histogram axes. The .mlx file made it easier for me to goes through the data in the RSVP data analysis code. CSO